Changed implementation from OpenCVCamera
so that it uses DirectShow as a default backend which solves the problem of delay opening the video stream. For the systems in which it is not available (as it is more or less deprecated right now), it falls back to Microsfot Media Foundation backend, which is the recommended on Windows. This also makes possible to use Droid Cam as input source.