Little refactor
I've removed the get_state thing as iews should not interact with the presenter. The update of the shortcuts is made via the already existing notify flow (no need to create other methods for this).

I've also tested the QDebug include and it builds without it, so I've removed it.