Fix initialization of camera_matrix using fov (field of view)
swap x_scale and y_scale in head3dScale since first row of transform involves pitch while 2nd row involves yaw.
simplify calculation of focalLength_width and focalLength_height for camera_matrix using known values of field of view, width and height.
We can calculate the diagonal of the triangle.
From the diagnoal, we can calculate the field of view for the width and height since they are the same ratio as the width and height as the diagonal..
From the field of view for width and height, we can calculate the focal length for the width and height since .
Note that focalLength_width is derived from fov_w which is derived from width.
focalLength_height is derived from fov_h which is derived from height.
Since the first row involves pitch, it is intialized with focalLength_height and height/2. The 2nd row involves yaw, so it is initialized with focalLength_width and width/2.  note the solvePnP API documentation indicates the first row involves x-axis and yaw while the 2nd row involves the y-axis and pitch indicates the implementation in PositionSolver swapped them