Fix yaw and pitch jitter at yaw and pitch boundary conditions
In correct_rotation, atan is the inverse function of tan canceling each other.
Note that lateral_offset/distance and vertical_offset/distance are already tangent ratios which require atan to obtain the angle in radians which is then multiplied by TO_DEG to obtain degrees.
After the corrections for yaw and pitch are added to the face_data rotations, we need to limit the result to +/-90.degrees since larger magnitudes are equivalent to facing away from the camera.
This fixes the large jitter in yaw and pitch at  yaw and pitch boundary conditions (+/- 90deg).