Fix init bug and some optimization
the ARRAY_SIZE is used for processing the array elements in a loop:
#define ARRAY_SIZE(_A) (sizeof(_A) / sizeof((_A)[0]))
However if a comma exists after the last element, there will be an extra row to the array containing all zeroes.
Declared some functions inline to improve performance since the are invoked within loops.
In ov534_reg_write, no need to copy the parameter val to a buffer before passing a pointer to libusb_control_transfer since the function is invoked repeatedly